verify token

+++ controllers/api/senior.js
@@ -16,17 +16,7 @@ module.exports.createSenior = async (req, res) => {
         req.body;
 
     try {
-        const token = req.cookies.access_token; // Get the token from cookies or authorization header
-        let owner;
-
-        if (token) {
-            const decoded = jwt.verify(token, process.env.JWT_SECRET); // Verify and decode the token
-            owner = decoded.id; // Assign the user ID as the owner
-        } else {
-            return res
-                .status(403)
-                .json({ description: 'Access denied. No token provided.' });
-        }
+        let owner = req.user.id;
 
         const newSenior = new Seniors({
             name,
@@ -56,6 +46,7 @@ module.exports.updateSenior = async (req, res) => {
     const { id } = req.params;
     const { name, domain, branch, year, whatsapp, telegram, college } =
         req.body;
+    console.log(req.body);
 
     try {
         const updatedSenior = await Seniors.findByIdAndUpdate(
@@ -76,6 +67,12 @@ module.exports.updateSenior = async (req, res) => {
             return res.status(404).json({ description: 'Senior not found' });
         }
 
+        if (updatedSenior.owner.toString() !== req.user.id) {
+            return res.status(403).json({
+                description: 'You are not authorized to Update this Senior',
+            });
+        }
+
         res.json({
             description: 'Senior updated successfully',
             updatedSenior,
@@ -86,17 +83,26 @@ module.exports.updateSenior = async (req, res) => {
     }
 };
 
-module.exports.deleteSenior = async (req, res, next) => {
+module.exports.deleteSenior = async (req, res) => {
     try {
-        const senior = await Seniors.findByIdAndDelete(req.params.id);
-
+        const senior = await Seniors.findById(req.params.id);
         if (!senior) {
             return res.status(404).json({ description: 'Senior not found' });
         }
 
-        res.status(200).json('Senior has been deleted...');
-    } catch (error) {
-        console.error('Error deleting Senior:', error);
+        if (senior.owner.toString() !== req.user.id) {
+            return res.status(403).json({
+                description: 'You are not authorized to delete this Senior',
+            });
+        }
+
+        await Seniors.findByIdAndDelete(req.params.id);
+
+        res.status(200).json({
+            description: 'Senior has been deleted successfully',
+        });
+    } catch (err) {
+        console.error('Error deleting Senior:', err);
         res.status(500).json({ description: 'Error deleting Senior' });
     }
 };
